A simple RSS feed display
I miss the iGoogle homepage, particularly the widgets that let you
include arbitrary RSS feeds as blocks on the page.

So I built this quick and simple replacement for making an RSS-driven
homepage for news.

Built in HTML, with JQuery, and the FeedEk and Masonry plugins.  Layout
is super simple and responsive, based on Bootstrap.  CDN versions are
referenced where possible.